pyproject.toml
[tool.poetry]
name = "weather"
version = "0.1.0"
description = ""
authors = ["alexandre-canet-E2S <121625353+alexandre-canet-E2S@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.2"
python-dotenv = "^1.0.0"
cdsapi = "^0.6.1"
xarray = "^2023.5.0"
e2slib = {git = "git@github.com:empowering-energy-solutions-ltd/e2slib.git", rev = "main"}
pvlib = "0.10.0"

[tool.poetry.group.dev.dependencies]
yapf = "^0.33.0"
pytest = "^7.3.1"
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
